# docs/swagger.yaml
basePath: /api/
definitions:
...
  httputil.HTTPError:
    properties:
      code:
        type: integer
      message:
        type: string
    type: object
...
  models.CreateUser:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.ShowPlayer:
    properties:
      age:
        example: 25
        type: integer
      country:
        example: Germany
        type: string
      first_name:
        example: Audrey
        type: string
      id:
        type: integer
      last_name:
        example: Hepburn
        type: string
      market_value:
        example: 25000
        type: integer
      position:
        example: 1
        type: integer
    type: object
  models.ShowTeam:
    properties:
      budget:
        type: integer
      country:
        type: string
      id:
        type: integer
      market_value:
        type: integer
      name:
        type: string
      players:
        items:
          $ref: '#/definitions/models.ShowPlayer'
        type: array
    type: object
...
  models.ShowUser:
    properties:
      email:
        type: string
      id:
        type: integer
      team:
        $ref: '#/definitions/models.ShowTeam'
    type: object
  models.UpdatePlayer:
    properties:
      age:
        example: 25
        type: integer
      country:
        example: Germany
        type: string
      first_name:
        example: Audrey
        type: string
      last_name:
        example: Hepburn
        type: string
      market_value:
        example: 25000
        type: integer
      position:
        example: 1
        type: integer
      team:
        type: integer
    type: object
  models.UpdateTeam:
    properties:
      budget:
        type: integer
      country:
        type: string
      name:
        type: string
      owner:
        type: integer
    type: object
  models.UpdateUser:
    properties:
      email:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Fantasy football manager microservice.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Fantasy football manager API
  version: "1.0"
paths:
...
  /teams:
    post:
      consumes:
      - application/json
      description: Create a new team
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth: []
      summary: Post a team
      tags:
      - Teams
...
  /users:
    post:
      consumes:
      - application/json
      description: Register a new user
      parameters:
      - description: CreateUser
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.CreateUser'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth: []
      summary: Create a user
      tags:
      - Users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID and all of it's associated resources
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth:
        - write
        - admin
      summary: Delete a user and all of it's associated resources
      tags:
      - Users
    get:
      consumes:
      - application/json
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ShowUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      summary: Get a user
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Update user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: UpdateUser data
        in: body
        name: email
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUser'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth:
        - admin
      summary: Update a user
      tags:
      - Users
  /users/me:
    get:
      consumes:
      - application/json
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ShowUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth:
        - write
        - admin
      summary: Get the logged in user
      tags:
      - Users
  /users/me/team:
    get:
      consumes:
      - application/json
      description: Get the logged in user's team
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ShowUser'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth: []
      summary: Get the logged in user's team
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Edit the logged in user's team
      parameters:
      - description: Update team payload
        in: body
        name: team
        required: true
        schema:
          $ref: '#/definitions/models.UpdateTeam'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth: []
      summary: Edit the logged in user's team
      tags:
      - Users
  /users/me/team/players:
    get:
      consumes:
      - application/json
      description: Get the logged in user's team players
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ShowPlayer'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth: []
      summary: Get the logged in user's team players
      tags:
      - Users
  /users/me/team/players/{id}:
    get:
      consumes:
      - application/json
      description: Get the logged in user's team player
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ShowPlayer'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth: []
      summary: Get the logged in user's team player
      tags:
      - Users
    patch:
      consumes:
      - application/json
      description: Get the logged in user's team players
      parameters:
      - description: Update player
        in: body
        name: player
        required: true
        schema:
          $ref: '#/definitions/models.UpdatePlayer'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/httputil.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/httputil.HTTPError'
      security:
      - BearerAuth: []
      summary: Edit the logged in user's team players
      tags:
      - Users
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"